project/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
SQLAlchemy
pymysql

langchain
langchain-openai
langchain-community
chromadb
openai
tiktoken

faker
graphviz
pydot

streamlit
requests

